refactor: extract parseDualModeArgs from duplicated disambiguation logic
Both parseSpanListArgs (span/list.ts) and parseLogListArgs (log/list.ts)
had near-identical control flow for detecting whether positional args
contain a trace ID. Extracted the shared logic into parseDualModeArgs()
in trace-target.ts, parameterized by the usage hint string.

Both commands now delegate to this shared utility via thin wrappers.

Addresses BugBot review feedback about duplicated argument disambiguation.
